% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_function.R
\name{create_bar_plot}
\alias{create_bar_plot}
\title{Creates a bar plot}
\usage{
create_bar_plot(
  data,
  x_var,
  y_var,
  title = "Bar Plot",
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{len}{Length of the object}

\item{width}{Width of the object}

\item{height}{Height of the object}
}
\value{
density The density of the object (numeric)
}
\description{
Use the length, width, height of an object to calculate its density.
Returns density as a floating point number.
}
