% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_function.R
\name{perform_anova_and_lm}
\alias{perform_anova_and_lm}
\title{Omit the na from data set}
\usage{
perform_anova_and_lm(data, dependent_var, independent_var)
}
\arguments{
\item{len}{Length of the object}

\item{width}{Width of the object}

\item{height}{Height of the object}
}
\value{
density The density of the object (numeric)
}
\description{
Use the length, width, height of an object to calculate its density.
Returns density as a floating point number.
}
